+++ Server/routes/meetups.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const routesMeetup = express.Router();
-const meetupsController  = require('../controllers/meetupsCotroller');
+const meetupsController  = require('../controllers/meetups');
 /*---------------------------------------------------------------------------- */
 /* Meetups End Points
 ================================================================================= */
@@ -8,10 +8,7 @@ const meetupsController  = require('../controllers/meetupsCotroller');
 routesMeetup.route('/api/v1/meetups')
 	.get(meetupsController.getAllMeetups)
 	.post(meetupsController.postAMeetup);
-routesMeetup.route('/api/v1/meetups/:id/rsvps')
-	.post(meetupsController.postARsvp);
-routesMeetup.route('/api/v1/meetups/upcoming')
-	.get(meetupsController.getAllUpcomingMeetups);
+	
 routesMeetup.route('/api/v1/meetups/:id')
 	.get(meetupsController.getAMeetup)
 	.put(meetupsController.updateAMeetup)
